+++ Help/Graph/jmcc-phase-modulation.hs
@@ -11,6 +11,6 @@ let nrec n f st = if n > 0 then nrec (n - 1) f (f st) else st
     y = mouseY kr 0 2 Linear 0.2 {- modulation index -}
     o (e,a) = let f = randId e 0 x
               in (succ e,fSinOsc ar (mce [f,f + rand2Id 'α' 1]) 0 * y + a)
-    ph = snd (nrec (3::Int) o ('β',0))
+    ph = snd (nrec 3 o ('β',0))
     freq = randId 'γ' 0 x
 in sinOsc ar (mce [freq, freq + rand2Id 'δ' 1]) ph * 0.1
